Moved database handle block in index file to database folder
