+++ config/env.js
@@ -2,4 +2,4 @@ import { config } from "dotenv";
 
 config({ path: `.env.${process.env.NODE_ENV || "development"}.local` });
 
-export const { PORT, NODE_ENV } = process.env;
+export const { PORT, NODE_ENV, DB_URI } = process.env;
