+++ database/mongodb.js
@@ -1,0 +1,18 @@
+import mongoose from "mongoose";
+import { NODE_ENV, DB_URI } from "../config/env.js";
+
+if (!DB_URI) {
+  throw new Error("Database URI is not defined in environment variables");
+}
+
+export const connectDatabase = async () => {
+  try {
+    await mongoose.connect(DB_URI);
+    console.log(`Connected to database in ${NODE_ENV} mode`);
+  } catch (error) {
+    console.error("Database connection failed:", error);
+    process.exit(1); // Exit the process with failure
+  }
+};
+
+// export default connectDatabase;
